some changes to itvs

+++ 4FS/ITVS/ITVS.X/input.h
@@ -12,21 +12,25 @@
 extern "C" {
 #endif
 
-#include <stdbool.h>
-#include <stdio.h>
-#include "mcc_generated_files/pin_manager.h"
-#include <xc.h>
-#include "onewire.h"
 #include "mcc_generated_files/mcc.h"
 
+
+/*** prototypes ***************************************************************/    
+//Fragt einen der 4 Inputs an und
+// liefert den aktuellen Zustand des Inputs zurrück
 bool GetInput(int in, bool out0, bool out1, bool out2, bool out3); 
 
+//sendet einen string via EUSART_Write
 void send_string(const char *x);
 
-float getTemp(void);
+//Fragt den Temperatursensor ab und liefert die Tempteratur 
+//als integer ein ein Zehntel Grad genau. (222 entsprechen 22,5 grad celsius))
+int getTemp(void);
+
 
 extern unsigned char SENSOR[9];
 
+//deklaration der Input Variablen
 bool out0 = false;
 bool out1 = false;
 bool out2 = false;
